De-duplicate the three config pages (rule of three)
Now that Roles is the third plugin config page, extract the pieces the
controllers and ops were copy-pasting:
- Ops::PluginConfiguration mixin: staged_activation + messages, with an
  abstract plugin_key each Configure op supplies.
- ConfiguresPlugin controller concern: plugin_enabled? + flash_for, keyed on
  controller_name (which already maps to each plugin).
- ChannelOptions presenter: the server-channel Tom Select option list the
  three config forms each rebuilt.

Skipped the role_sets position unique index Fable suggested: positions are
unique by construction (always max+1), so the index would be defensive
against a state we don't create.
