+++ app/components/logging/config_form.rb
@@ -145,8 +145,6 @@ class Components::Logging::ConfigForm < Components::Base
   end
 
   def channels
-    @channels ||= @config.server_channels.text.map do |channel|
-      Components::TomSelect::Option.for(value: channel.discord_id, label: channel.name)
-    end
+    @channels ||= ChannelOptions.new(@config).options
   end
 end
