+++ app/components/welcomes/config_form.rb
@@ -130,9 +130,7 @@ class Components::Welcomes::ConfigForm < Components::Base
   end
 
   def channels
-    @channels ||= @config.server_channels.text.map do |channel|
-      Components::TomSelect::Option.for(value: channel.discord_id, label: channel.name)
-    end
+    @channels ||= ChannelOptions.new(@config).options
   end
 
   def camel(name)
