+++ app/operations/logging/configure.rb
@@ -3,6 +3,8 @@
 module Ops
   module Logging
     class Configure < ApplicationOperation
+      include Ops::PluginConfiguration
+
       receives :server_configuration, :channel_id, :enabled_actions, :enabled
 
       def call
@@ -19,14 +21,8 @@ module Ops
 
       private
 
-      def staged_activation
-        activation = server_configuration.plugin_activations.find_or_initialize_by(plugin: Plugin.find_by!(key: :logging))
-        activation.enabled = enabled
-        activation
-      end
-
-      def messages(*records)
-        records.flat_map { |record| record.errors.full_messages }
+      def plugin_key
+        :logging
       end
     end
   end
